Replace hypen with underscores for tests folders (#377)
Replaced hypens with underscores for tests-system, tests-unit and
tests-integration folders and relevant changes to their references in
files.

In the `test_runner.py` files need to replace the `underscores` with
`hyphens` for `_tool_main_path` variable as this is used to execute the
tool and the tool names contains `hyphens` instead of `underscores`.
**This will be reverted back if we decide to rename the lobster tools
with underscores.**

Reason for change - Each folder should be importable as a Python module
and hence we need to replace `-` with `_` in the folder names.