Fix tracing reports and delete old format Python system test (#380)
Due to the replacement of `-` with `_` in `tests_system` &
`tests-integration` the tracing target in Makefile was failing to
generate the tracing reports.

**Issue**
The issue was that the `unit tests` and `system tests` targets were
looking for folder like `tests-unit/lobster-tool-name` &
`tests-system/lobster-tool-name` but instead of hyphens it should be
underscores in the folders. This PR fixes that and the generation of the
reports.

**NOTE**
This PR also deletes the system tests for `lobster-python` which will be
upgraded to new STF and is tracked in [this
ticket](https://jira.cc.bmwgroup.net/browse/SWF-20309).