Add Type Hierarchy System Test for `lobster-trlc` (#390)
Add a system test that verifies the behavior of `lobster-trlc` with
respect to extended TRLC types.

In the test input data two TRLC record types exist:
- a base type
- and a type which extends the base type

Depending on the configuration of `lobster-trlc` either all TRLC record
objects of both types shall be converted to LOBSTER objects, or only the
extended type.

Converting only the base type is not supported.